Chuyển lenght, width, height sang double. Sửa lỗi Exception khi từ chối. Sửa lỗi set default media

+++ backend/src/main/java/com/lumina_book/backend/dto/request/ProductUpdateRequest.java
@@ -33,13 +33,13 @@ public class ProductUpdateRequest {
     Double weight;
 
     @Min(value = 1, message = "Chiều dài phải lớn hơn hoặc bằng 1")
-    Integer length;
+    Double length;
 
     @Min(value = 1, message = "Chiều rộng phải lớn hơn hoặc bằng 1")
-    Integer width;
+    Double width;
 
     @Min(value = 1, message = "Chiều cao phải lớn hơn hoặc bằng 1")
-    Integer height;
+    Double height;
 
     @DecimalMin(value = "0.0", message = "Giá sản phẩm phải lớn hơn hoặc bằng 0")
     Double price;
